requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coverage==5.5
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cleanup==5.1.0
django-cors-headers==3.7.0
django-dbbackup==3.3.0
django-debug-toolbar==3.2
django-filter==2.4.0
django-import-export==2.5.0
django-jazzmin==2.4.5
django-prometheus==2.1.0
django-timezone-field==4.1.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
et-xmlfile==1.0.1
flake8==3.9.0
future==0.18.2
jdcal==1.4.1
kombu==5.0.2
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.2.0
pipdeptree==2.0.0
prometheus-client==0.10.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
